+++ Src/Web/Web.Shared.Net/RequestTrackingTelemetryModule.cs
@@ -20,13 +20,14 @@
     public class RequestTrackingTelemetryModule : ITelemetryModule
     {
         private const string IntermediateRequestHttpContextKey = "IntermediateRequest";
-        private readonly IList<string> handlersToFilter = new List<string>();
+        // if HttpApplicaiton.OnRequestExecute is available, we don't attempt to detect any correlation issues
+        private static bool correlationIssuesDetectionComplete = typeof(HttpApplication).GetMethod("OnExecuteRequestStep") != null;
+
         private TelemetryClient telemetryClient;
         private TelemetryConfiguration telemetryConfiguration;
         private bool initializationErrorReported;
-        private bool correlationHeadersEnabled = true;
         private ChildRequestTrackingSuppressionModule childRequestTrackingSuppressionModule = null;
-        
+
         /// <summary>
         /// Gets or sets a value indicating whether child request suppression is enabled or disabled. 
         /// True by default.
@@ -53,34 +54,17 @@
         /// See also <see cref="ChildRequestTrackingSuppressionModule" />.
         /// </remarks>
         public int ChildRequestTrackingInternalDictionarySize { get; set; }
-        
+
         /// <summary>
         /// Gets the list of handler types for which requests telemetry will not be collected
         /// if request was successful.
         /// </summary>
-        public IList<string> Handlers
-        {
-            get
-            {
-                return this.handlersToFilter;
-            }
-        }
+        public IList<string> Handlers { get; } = new List<string>();
 
         /// <summary>
         /// Gets or sets a value indicating whether the component correlation headers would be set on http responses.
         /// </summary>
-        public bool SetComponentCorrelationHttpHeaders
-        {
-            get
-            {
-                return this.correlationHeadersEnabled;
-            }
-
-            set
-            {
-                this.correlationHeadersEnabled = value;
-            }
-        }
+        public bool SetComponentCorrelationHttpHeaders { get; set; } = true;
 
         /// <summary>
         /// Gets or sets the endpoint that is to be used to get the application insights resource's profile (appId etc.).
@@ -219,6 +203,12 @@
                 }
 
                 this.telemetryClient.TrackRequest(requestTelemetry);
+
+                if (!correlationIssuesDetectionComplete && context.Request.ContentLength > 0)
+                {
+                    WebEventSource.Log.CorrelationIssueIsDetectedForRequestWithBody();
+                    correlationIssuesDetectionComplete = true;
+                }
             }
             else
             {
@@ -394,7 +384,7 @@
 
             return false;
         }
-        
+
         /// <summary>
         /// <see cref="System.Web.Handlers.TransferRequestHandler"/> can create a Child request to route extension-less requests to a controller.
         /// (ex: site/home -> site/HomeController.cs)