+++ src/Microsoft.ApplicationInsights.AspNetCore/Extensions/HttpRequestExtensions.cs
@@ -9,13 +9,16 @@
     /// </summary>
     public static class HttpRequestExtensions
     {
+
+        private const string UnknownHostName = "UNKNOWN-HOST";
+
         /// <summary>
         /// Gets http request Uri from request object
         /// </summary>
         /// <param name="request">The <see cref="HttpRequest"/></param>
         /// <returns>A New Uri object representing request Uri</returns>
         public static Uri GetUri(this HttpRequest request)
-        {
+        {            
             if (null == request)
             {
                 throw new ArgumentNullException("request");
@@ -26,16 +29,13 @@
                 throw new ArgumentException("Http request Scheme is not specified");
             }
 
-            if (false == request.Host.HasValue)
-            {
-                throw new ArgumentException("Http request Host is not specified");
-            }
-
+            string hostName = request.Host.HasValue ? request.Host.ToString() : UnknownHostName;
+            
             var builder = new StringBuilder();
 
             builder.Append(request.Scheme)
                 .Append("://")
-                .Append(request.Host);
+                .Append(hostName);
 
             if (true == request.Path.HasValue)
             {